WIP migrating test_adapted_attributes
tests/test_adapted_attributes.py::test_adapted_filepath_type throws datajoint/s3.py:54: BucketInaccessible
